:recycle: Refactor: 애러 코드 위치 이동 (#5)
